Update from BZR repo

+++ mysql-test/mutlib/mutlib.py
@@ -1,6 +1,5 @@
-#!/usr/bin/env python
 #
-# Copyright (c) 2010, 2012 Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2010, 2013 Oracle and/or its affiliates. All rights reserved.
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -29,9 +28,16 @@ import difflib
 import os
 import string
 import subprocess
-import sys
-import time
+import tempfile
+
+from mysql.utilities.common.my_print_defaults import MyDefaultsReader
+from mysql.utilities.common.my_print_defaults import my_login_config_path
+from mysql.utilities.common.tools import get_tool_path
+
 from mysql.utilities.exception import MUTLibError
+from mysql.utilities.exception import UtilDBError
+from mysql.utilities.exception import UtilError
+
 
 # Constants
 MAX_SERVER_POOL = 10
@@ -188,7 +194,7 @@ class Server_list(object):
         full_datadir = os.getcwd() + "/temp_%s" % port
         
         # Attempt to clone existing server
-        cmd = "mysqlserverclone.py --server="
+        cmd = "mysqlserverclone.py --delete-data --server="
         cmd += self.get_connection_string(cur_server)
         if passwd:
            cmd += " --root-password=%s " % passwd
@@ -222,7 +228,7 @@ class Server_list(object):
         # Connect to the new instance
         try:
             self.new_server.connect()
-        except MUTLibError, e:
+        except UtilError, e:
             raise MUTLibError("Cannot connect to spawned server: %s" % \
                                e.errmsg)
             
@@ -266,11 +272,36 @@ class Server_list(object):
         # Build the shutdown command
         cmd = ""
         res = server.show_server_variable("basedir")
-        cmd += os.path.normpath(os.path.join(res[0][1], "bin", "mysqladmin"))
+        mysqladmin_client = "mysqladmin"
+        if not os.name == "posix":
+            mysqladmin_client = "mysqladmin.exe"
+        mysqladmin_path= os.path.normpath(os.path.join(res[0][1], "bin",
+                                                   mysqladmin_client))
+        if not os.path.exists(mysqladmin_path):
+            mysqladmin_path= os.path.normpath(os.path.join(res[0][1], "client",
+                                                       mysqladmin_client))
+        if not os.path.exists(mysqladmin_path) and not os.name == 'posix':
+            mysqladmin_path= os.path.normpath(os.path.join(res[0][1],
+                                                       "client/debug",
+                                                       mysqladmin_client))
+        if not os.path.exists(mysqladmin_path) and not os.name == 'posix':
+            mysqladmin_path= os.path.normpath(os.path.join(res[0][1],
+                                                       "client/release",
+                                                       mysqladmin_client))
+        cmd += mysqladmin_path
         cmd += " shutdown "
         cmd += self.get_connection_parameters(server)
         res = server.show_server_variable("datadir")
         datadir = res[0][1]
+
+        # Kill all connections so shutdown will work correctly
+        res = server.exec_query("SHOW PROCESSLIST")
+        for row in res:
+            if not row[7] or not row[7].upper().startswith("SHOW PROCESS"):
+                try:
+                    server.exec_query("KILL CONNECTION %s" % row[0])
+                except UtilDBError: # Ok to ignore KILL failures
+                    pass
 
         # disconnect user
         server.disconnect()
@@ -278,7 +309,7 @@ class Server_list(object):
         # Stop the server
         file = os.devnull
         f_out = open(file, 'w')
-        proc = subprocess.Popen(cmd, shell=True,
+        proc = subprocess.Popen(cmd, shell=True, 
                                 stdout = f_out, stderr = f_out)
         ret_val = proc.wait()
         f_out.close()
@@ -316,12 +347,11 @@ class Server_list(object):
         
         cur_num_servers = self.num_servers()
         for server_num in range(0, num_to_add):
-            datadir = "new_server_%d" % (cur_num_servers)
             cur_num_servers += 1
             server = self.start_new_server(orig_server,
                                            self.get_next_port(),
-                                           self.get_next_id(), "root",
-                                           datadir)
+                                           self.get_next_id(), "root")
+            datadir = server[0].show_server_variable('datadir')[0][1]
             self.server_list.append((server[0], True,
                                      self.get_process_id(datadir)))
 
@@ -356,15 +386,15 @@ class Server_list(object):
                     print "  Shutting down server %s..." % server[0].role,
                     if self.stop_server(server[0]):
                         print "success."
-                    elif server[2] is not None:
+                    elif server[2] is not None and server[2] > 1:
                         print "WARN - attempting SIGTERM - pid = %s" % server[2] 
                         # try signal termination
                         if os.name == "posix":
-                            retval = os.kill(server[2],
+                            retval = os.kill(int(server[2]),
                                              subprocess.signal.SIGTERM)
                         else:
                             retval = subprocess.Popen("taskkill /F /T /PID %i" %
-                                                      server[2],
+                                                      int(server[2]),
                                                       shell=True)
                     else:
                         print "ERROR"
@@ -381,6 +411,9 @@ class Server_list(object):
         id[in]             The process id if known
         """
         if new_server is not None:
+            if id == -1:
+                datadir = new_server.show_server_variable("datadir")[0][1]
+                id = self.get_process_id(datadir)
             self.server_list.append((new_server, spawned, id))
 
 
@@ -583,6 +616,89 @@ class System_test(object):
 
         return False
 
+    def check_mylogin_requisites(self):
+        """ Check if the tools to manipulate mylogin.cnf are accessible.
+
+        This method verifies if the MySQL client tools my_print_defaults and
+        mysql_config_editor are accessible.
+
+        A MUTLibError exception is raised if the requisites are not met.
+        """
+        try:
+            self.login_reader = MyDefaultsReader(
+                                    find_my_print_defaults_tool=True)
+        except UtilError as err:
+            raise MUTLibError("MySQL client tools must be accessible to run "
+                              "this test (%s). E.g. Add the location of the "
+                              "MySQL client tools to your PATH." % err.errmsg)
+
+        if not self.login_reader.check_login_path_support():
+            raise MUTLibError("ERROR: the used my_print_defaults tool does not "
+                            "support login-path options. Used tool: %s"
+                            % self.login_reader.tool_path)
+
+        try:
+            self.edit_tool_path = get_tool_path(None, "mysql_config_editor",
+                                                search_PATH=True)
+        except UtilError as err:
+            raise MUTLibError("MySQL client tools must be accessible to run "
+                              "this test (%s). E.g. Add the location of the "
+                              "MySQL client tools to your PATH." % err.errmsg)
+
+    def create_login_path_data(self, login_path, user, host):
+        """Add the specified login-path data to .mylogin.cnf.
+
+        Execute mysql_config_editor tool to create a new login-path
+        entry to the .mylogin.cnf file.
+
+        Note: the use of password is not supported because it is not read from
+        the stdin by the tool (apparently for security reasons).
+        """
+
+        assert self.edit_tool_path, ("The tool mysql_config_editor is not "
+                                     "accessible. First, use method "
+                                     "check_mylogin_requisites.")
+
+        cmd = [self.edit_tool_path]
+        cmd.append('set')
+        cmd.append('--login-path=%s' % login_path)
+        cmd.append('--host=%s' % host)
+        cmd.append('--user=%s' % user)
+
+        # Create a temporary file to redirect stdout
+        out_file = tempfile.TemporaryFile()
+
+        # Execute command to create login-path data
+        proc = subprocess.Popen(cmd, stdout=out_file,
+                                stdin=subprocess.PIPE)
+        # Overwrite login-path if already exists (i.e. answer 'y' to question)
+        proc.communicate('y')
+
+    def remove_login_path_data(self, login_path):
+        """Remove the specified login-path data from .mylogin.cnf.
+
+        Execute mysql_config_editor tool to remove the specified login-path
+        entry from the .mylogin.cnf file.
+        """
+        assert self.edit_tool_path, ("The tool mysql_config_editor is not "
+                                     "accessible. First, use method "
+                                     "check_mylogin_requisites.")
+
+        cmd = [self.edit_tool_path]
+        cmd.append('remove')
+        cmd.append('--login-path=%s' % login_path)
+
+        # Create a temporary file to redirect stdout
+        out_file = tempfile.TemporaryFile()
+
+        # Execute command to remove login-path data
+        if self.verbose:
+            subprocess.call(cmd, stdout=out_file)
+        else:
+            # Redirect stderr to null
+            null_file = open(os.devnull, "w+b")
+            subprocess.call(cmd, stdout=out_file,
+                            stderr=null_file)
 
     def exec_util(self, cmd, file_out, abspath=False):
         """Execute Utility
@@ -739,7 +855,28 @@ class System_test(object):
         for linenum in range(len(linenums)-1, -1, -1):
             self.results.pop(linenums[linenum])
 
+    def remove_result_and_lines_before(self, prefix, lines=1):
+        """Remove lines in the results.
     
+        prefix[in]         starting prefix of string to mask
+        lines[in]          number of lines to remove previously
+                           to the prefix line.
+        """
+        linenums = []
+        linenum = 0
+        for line in self.results:
+            index = line.find(prefix)
+            if index == 0:
+                linenums.append(int(linenum))
+                for line2rm in range(linenum-lines,linenum):
+                    if line2rm > - 1:
+                        linenums.append(int(line2rm))
+            linenum += 1
+        linenums.sort()
+        # Must remove lines in reverse order
+        for linenum in range(len(linenums) - 1, - 1, - 1):
+            self.results.pop(linenums[linenum])
+
     def replace_substring(self, target, replacement):
         """Replace a target substring in the entire result file.
